Extract Radar drawing responsibility out of Radar
